+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Next version
 
+- Upgrade to @mangrovedao/mangrove-deployments v2.1.1
+- Upgrade to @mangrovedao/context-addresses v1.2.0
+
 # 2.1.0-0
 
 - Upgrade to @mangrovedao/mangrove-deployments v2.0.1