+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Next version
 
 - Upgrade to @mangrovedao/mangrove-deployments v2.1.1
-- Upgrade to @mangrovedao/context-addresses v1.2.2-1
+- Upgrade to @mangrovedao/context-addresses v1.3.0
 
 ## Blast support